+++ spring-boot-project/spring-boot-docker-compose/src/main/java/org/springframework/boot/docker/compose/lifecycle/DockerComposeProperties.java
@@ -75,6 +75,8 @@ public class DockerComposeProperties {
 
 	private final Skip skip = new Skip();
 
+	private final Readiness readiness = new Readiness();
+
 	public boolean isEnabled() {
 		return this.enabled;
 	}
@@ -121,6 +123,10 @@ public class DockerComposeProperties {
 
 	public Skip getSkip() {
 		return this.skip;
+	}
+
+	public Readiness getReadiness() {
+		return this.readiness;
 	}
 
 	static DockerComposeProperties get(Binder binder) {
@@ -229,6 +235,68 @@ public class DockerComposeProperties {
 
 		public void setInTests(boolean inTests) {
 			this.inTests = inTests;
+		}
+
+	}
+
+	/**
+	 * Readiness properties.
+	 */
+	public static class Readiness {
+
+		/**
+		 * Timeout of the readiness checks.
+		 */
+		private Duration timeout = Duration.ofMinutes(2);
+
+		/**
+		 * TCP properties.
+		 */
+		private final Tcp tcp = new Tcp();
+
+		public Duration getTimeout() {
+			return this.timeout;
+		}
+
+		public void setTimeout(Duration timeout) {
+			this.timeout = timeout;
+		}
+
+		public Tcp getTcp() {
+			return this.tcp;
+		}
+
+		/**
+		 * TCP properties.
+		 */
+		public static class Tcp {
+
+			/**
+			 * Timeout for connections.
+			 */
+			private Duration connectTimeout = Duration.ofMillis(200);
+
+			/**
+			 * Timeout for reads.
+			 */
+			private Duration readTimeout = Duration.ofMillis(200);
+
+			public Duration getConnectTimeout() {
+				return this.connectTimeout;
+			}
+
+			public void setConnectTimeout(Duration connectTimeout) {
+				this.connectTimeout = connectTimeout;
+			}
+
+			public Duration getReadTimeout() {
+				return this.readTimeout;
+			}
+
+			public void setReadTimeout(Duration readTimeout) {
+				this.readTimeout = readTimeout;
+			}
+
 		}
 
 	}
