+++ spring-boot-project/spring-boot-docker-compose/src/test/java/org/springframework/boot/docker/compose/lifecycle/DockerComposePropertiesTests.java
@@ -48,6 +48,9 @@ class DockerComposePropertiesTests {
 		assertThat(properties.getStop().getCommand()).isEqualTo(StopCommand.STOP);
 		assertThat(properties.getStop().getTimeout()).isEqualTo(Duration.ofSeconds(10));
 		assertThat(properties.getProfiles().getActive()).isEmpty();
+		assertThat(properties.getReadiness().getTimeout()).isEqualTo(Duration.ofMinutes(2));
+		assertThat(properties.getReadiness().getTcp().getConnectTimeout()).isEqualTo(Duration.ofMillis(200));
+		assertThat(properties.getReadiness().getTcp().getReadTimeout()).isEqualTo(Duration.ofMillis(200));
 	}
 
 	@Test
@@ -60,6 +63,9 @@ class DockerComposePropertiesTests {
 		source.put("spring.docker.compose.stop.command", "down");
 		source.put("spring.docker.compose.stop.timeout", "5s");
 		source.put("spring.docker.compose.profiles.active", "myprofile");
+		source.put("spring.docker.compose.readiness.timeout", "10s");
+		source.put("spring.docker.compose.readiness.tcp.connect-timeout", "400ms");
+		source.put("spring.docker.compose.readiness.tcp.read-timeout", "500ms");
 		Binder binder = new Binder(new MapConfigurationPropertySource(source));
 		DockerComposeProperties properties = DockerComposeProperties.get(binder);
 		assertThat(properties.getFile()).isEqualTo(new File("my-compose.yml"));
@@ -69,6 +75,9 @@ class DockerComposePropertiesTests {
 		assertThat(properties.getStop().getCommand()).isEqualTo(StopCommand.DOWN);
 		assertThat(properties.getStop().getTimeout()).isEqualTo(Duration.ofSeconds(5));
 		assertThat(properties.getProfiles().getActive()).containsExactly("myprofile");
+		assertThat(properties.getReadiness().getTimeout()).isEqualTo(Duration.ofSeconds(10));
+		assertThat(properties.getReadiness().getTcp().getConnectTimeout()).isEqualTo(Duration.ofMillis(400));
+		assertThat(properties.getReadiness().getTcp().getReadTimeout()).isEqualTo(Duration.ofMillis(500));
 	}
 
 }
